Refactor terms of service content for use across multiple components
Add new route for undecorated (static) terms page shown in mobile apps.
